+++ src/commands/tokens/send_ft/mod.rs
@@ -2,6 +2,7 @@ use strum::{EnumDiscriminants, EnumIter, EnumMessage};
 
 mod exact_amount_ft;
 mod max_amount_ft;
+mod transaction_formation;
 
 #[derive(Debug, Clone, interactive_clap::InteractiveClap)]
 #[interactive_clap(input_context = super::TokensCommandsContext)]
